+++ NEWS.md
@@ -1,5 +1,6 @@
 Version 0.9.1 For Dev-C++ 7 Beta
  - enhancement: code completion suggestion for "__func__" variable
+ - fix: ide failed to start, if there are errors in the compiler set settings
 
 Version 0.9.0 For Dev-C++ 7 Beta
  - fix: control keys in the numpad doesn't work in the editor